Idea for Engine
This branch represents my first idea for the way that the main engine in videogame might work. It has a class representing the engine itself which keeps track of the console, externalized strings, externalized screens, the objects, and the interactions between these objects.
This engine may completely differ from the engine eventually used in videogame, and that's ok.