This branch takes a different approach.
Instead of incorporating the user input into the screen, making the input gathering a separate object entirely.

This branch is a response to the inexplicable not-applying-globally of the input that the serialized-variables branch is experiencing. I can only assume that that issue is some strange oddity with how I externalized the screens.